fix[menu]: menu asset implementation
[Description]:
This change effectively introduces a reworked main menu with a set of
new compatible animations that look much nicer than the previous
animation. The current animation still needs to have a bit of work with
the squash and stretch stuff because it still looks a bit rigid but
overall its a much better impelementation now.

[Breaking Changes]:
None :)

[Related Issues/PRs]:
closes #2

[Tags/Labels]:
UI