feat(docs): web documentation page with Sphinx
[Description]:
Generated a brief installation page for the main repo. This is mainly
copied off the sphinx-awesome documentation theme.

[Breaking Changes]:
I don't think tox will break but there are definitely some developer
requirements that will need to be met (namely the ones in the
sphinx-awesome `requirements.txt` file) if tox is to run with `make
html` in the `docs` folder.

[Related Issues/PRs]:
ref: closes #13

[Tags/Labels]:
Enhancement
Documentation