feat: player class, bugfixes
[Description]:
Fixed gitignore somehow including .py in the gitignore. This is crazy,
but anyhow the conf.py file for the documentation is now able to pass
staging so i guess thats something.

The player dunder files are ready in the entities class which basically
has 0 functionality but all the properties and most of the physics are
there. the only things that need to be addressed are filling out the
physical methods themselves.

[Breaking Changes]:
No breaking changes noticed.

[Related Issues/PRs]:
ref: #18

[Tags/Labels]:
Feature
Bugfix
Bug
Feat
Enhancement