feat(engine): optimization, unit tests
[Description]:
This further optimizes the engine through the utilities provided by
pygame such as vectors.

Said vectors also allow for decent testing and readability when it comes
to unit testing.

Note: not all tests are done and only about a 5th of the actual tests
are done.

[Breaking Changes]:
None listed. Tests work.

[Co-authors]:
Co-authors of the commit can be added using "Co-authored-by" lines, to give credit to all those who contributed to the change.

Co-authored-by: DaFluffyPotato dafluffypotatodev@gmail.com