tests(unit): unit tests, tox changes
This commit addresses issues where tox would not work for some displays
and especially on server side testing where display / vm's are not
practical.

This implementation fixes that by skipping tests that do not have any
display in os.environ and justifies it with a reason.

In addition to these changes, the tox environment has also undergone
changes including reducing the amount of load that code blackening will
occur